@@ -28,23 +28,21 @@ export function reducer(state = initialState, action: BlogActionsUnion): State {
     case BlogActionTypes.LOAD_ALL_BLOGS:
     case BlogActionTypes.LOAD_ONE_BLOG:
     case BlogActionTypes.ADD_BLOG:
-    case BlogActionTypes.EDIT_BLOG:
-    case BlogActionTypes.SAVE_BLOG_DRAFT: {
+    case BlogActionTypes.EDIT_BLOG: {
       return {
         ...state,
         selectedBlogId: undefined,
         errorMessage: undefined,
-      }
+      };
     }
     case BlogActionTypes.LOAD_ALL_BLOGS_FAIL:
     case BlogActionTypes.LOAD_ONE_BLOG_FAIL:
     case BlogActionTypes.ADD_BLOG_FAIL:
-    case BlogActionTypes.EDIT_BLOG_FAIL:
-    case BlogActionTypes.SAVE_BLOG_DRAFT_FAIL: {
+    case BlogActionTypes.EDIT_BLOG_FAIL: {
       return {
         ...state,
         errorMessage: action.payload,
-      }
+      };
     }
     case BlogActionTypes.LOAD_ALL_BLOGS_SUCCESS: {
       return adapter.addAll(action.payload, { ...state, errorMessage: 'Success' });
@@ -67,9 +65,6 @@ export function reducer(state = initialState, action: BlogActionsUnion): State {
         ...state,
         errorMessage: 'Success',
       });
-    }
-    case BlogActionTypes.SAVE_BLOG_DRAFT_SUCCESS: {
-      
     }
     default: {
       return state;
